# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Cafe Manager API",
    "version": "1.0.0",
    "description": "REST API for the Smart Cafe Manager multi-tenant platform — Next.js App Router Route Handlers backed by Supabase. This spec is generated from `@swagger` JSDoc blocks on route handlers by scripts/generateSwagger.js; view it rendered at /api-docs. Only a representative subset of routes carries `@swagger` annotations so far (see the plan doc, Phase 11) — extend the pattern to the rest incrementally by adding the same style of block above any other exported GET/POST/PATCH/DELETE handler."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this app"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by /api/auth/staff-login, /api/auth/customer-login, or /api/auth/register. Send as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "hiveId": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "master_admin",
              "owner",
              "manager",
              "cook",
              "waiter",
              "customer"
            ]
          },
          "authProvider": {
            "type": "string"
          },
          "spaceId": {
            "type": "string",
            "nullable": true
          },
          "preferences": {
            "type": "object",
            "properties": {
              "orderCount": {
                "type": "integer"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/staff-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Staff/manager/owner/cook/waiter login",
        "description": "Password login for every non-customer role. Returns a JWT carrying userId, role, orgId, and isMasterAdmin claims.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier",
                  "password"
                ],
                "properties": {
                  "identifier": {
                    "type": "string",
                    "description": "Email or phone number",
                    "example": "owner@cafe.test"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Staff login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Identifier or password missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new customer, or log in an existing one",
        "description": "Customers never have a password, so finding an existing account by hiveId/email/phone IS the login path. name is only required when actually creating a new account.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "hiveId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Required only for new accounts"
                  },
                  "hiveId": {
                    "type": "string",
                    "description": "Unique id for this device/install"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "spaceId": {
                    "type": "string",
                    "description": "Optional space id resolved from a scanned QR code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing account logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "isNewUser": {
                      "type": "boolean"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields, or the QR code is no longer valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email or phone already in use by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/customer-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Returning customer login (passwordless)",
        "description": "Logs an existing customer back in on a new device/install by email or phone, moving their hiveId to this device. Optionally re-links them to a space via a scanned QR's spaceId.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifier",
                  "hiveId"
                ],
                "properties": {
                  "identifier": {
                    "type": "string",
                    "description": "Email or phone used at registration"
                  },
                  "hiveId": {
                    "type": "string",
                    "description": "Unique id for this device/install"
                  },
                  "spaceId": {
                    "type": "string",
                    "description": "Optional space id resolved from a scanned QR code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing identifier/hiveId, or the QR code is no longer valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No account found with that email or phone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}